cleaning up point chart

+++ PointChart/Common/DomainModel/Chart.cs
@@ -21,7 +21,7 @@ namespace AlwaysMoveForward.PointChart.Common.DomainModel
         /// <summary>
         /// Gets and sets the unique identifier of the chart
         /// </summary>
-        public int Id { get; set; }
+        public long Id { get; set; }
         
         /// <summary>
         /// Gets and sets name of the chart
@@ -31,7 +31,7 @@ namespace AlwaysMoveForward.PointChart.Common.DomainModel
         /// <summary>
         /// Gets and sets the identifier of the person that adminsters this chart/point earner
         /// </summary>
-        public int AdministratorId { get; set; }
+        public long AdministratorId { get; set; }
         
         /// <summary>
         /// Gets and sets all the tasks on the chart
